refactor events
