+++ client/app/header.js
@@ -7,15 +7,17 @@ import { logout } from '../auth/authActions';
 import { fetchTrips } from '../trip/tripActions';
 import { deselectTrip } from '../tripPlan/tripPlanActions';
 import { clearAll } from '../tripPlan/dest/destActions';
+import { clearDest } from '../tripPlan/event/eventActions';
 
 const mapDispatchToProps = dispatch => ({
   onLogout: id => dispatch(logout()),
   onClickTrips: () => dispatch(fetchTrips()),
   onClear: () => dispatch(clearAll()),
+  onClearEvents: () => dispatch(clearDest()),
   onDeselect: () => dispatch(deselectTrip()),
 });
 
-let Header = ({ onLogout, onClickTrips, onClear, onDeselect }) => (
+let Header = ({ onLogout, onClickTrips, onClear, onDeselect, onClearEvents }) => (
   <AppBar title="voyAger" showMenuIconButton={false}>
     <Link to="/trip"><Tab label="My Trips" onClick={ () =>
       onClickTrips() }
@@ -24,6 +26,7 @@ let Header = ({ onLogout, onClickTrips, onClear, onDeselect }) => (
     <Link to="/home"><Tab label="Home" onClick={ () => {
       onClear();
       onDeselect();
+      onClearEvents();
     }}
     />
     </Link>
@@ -35,6 +38,7 @@ Header.propTypes = {
   onLogout: React.PropTypes.func.isRequired,
   onClickTrips: React.PropTypes.func.isRequired,
   onClear: React.PropTypes.func.isRequired,
+  onClearEvents: React.PropTypes.func.isRequired,
   onDeselect: React.PropTypes.func.isRequired,
 };
