+++ client/tripPlan/dest/destReducers.js
@@ -1,11 +1,10 @@
 import Immutable from 'immutable';
 import { REQUEST_DESTINATIONS, RECEIVE_DESTINATIONS,
-FETCH_DEST_FAILURE, RECEIVE_EVENTS_IN_DEST } from './destActions';
+FETCH_DEST_FAILURE } from './destActions';
 
 const initialState = Immutable.Map({
   isFetching: false,
   destinations: Immutable.List(),
-  events: Immutable.List(),
   key: null,
 });
 
@@ -19,10 +18,6 @@ const destination = (state, action) => {
 
 const destinations = (state = initialState, action) => {
   switch (action.type) {
-    case RECEIVE_EVENTS_IN_DEST:
-      const oldEvents = state.get('events');
-      const newEvents = oldEvents.push(action.payload.events);
-      return state.set('events', newEvents);
     case REQUEST_DESTINATIONS:
       return state.set('isFetching', true);
     case RECEIVE_DESTINATIONS:
@@ -30,11 +25,8 @@ const destinations = (state = initialState, action) => {
       return state.set('destinations', action.payload.destinations);
     case 'CHOOSE_DEST':
       return state.set('key', action.payload.key);
-    case 'CLEAR_DEST':
-      return state.set('events', Immutable.List());
     case 'CLEAR_ALL':
       state = state.set('destinations', Immutable.List());
-      state = state.set('events', Immutable.List());
       state = state.set('key', null);
       return state;
     case FETCH_DEST_FAILURE:
