+++ client/tripPlan/event/eventReducers.js
@@ -1,13 +1,18 @@
 import Immutable from 'immutable';
+import { RECEIVE_EVENTS_IN_DEST } from './eventActions';
 
 const initialState = Immutable.Map({
-  key: null,
+  events: Immutable.List(),
 });
 
 export default (state = initialState, action) => {
   switch (action.type) {
-    case 'CHOOSE_KEY':
-      return state.set('key', action.payload.key);
+    case RECEIVE_EVENTS_IN_DEST:
+      const oldEvents = state.get('events');
+      const newEvents = oldEvents.push(action.payload.events);
+      return state.set('events', newEvents);
+    case 'CLEAR_DEST':
+      return state.set('events', Immutable.List());
     default:
       return state;
   }
